refactor GNI module and gniWorld reducer
