+++ package.json
@@ -17,6 +17,7 @@
     "redux": "^4.1.1",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.3.0",
+    "sass": "^1.39.0",
     "web-vitals": "^1.0.1"
   },
   "scripts": {
