+++ src/components/GroupsList/GroupsList.scss
@@ -1,0 +1,3 @@
+.cardBg {
+  background-color: pink;
+}
