+++ src/styles/abstracts/_fonts.scss
@@ -1,0 +1,2 @@
+$roboto-slab-font-family: 'Roboto Slab', serif;
+$montserrat-font-family: 'Montserrat', sans-serif;
