+++ src/styles/abstracts/_index.scss
@@ -1,0 +1,7 @@
+@forward 'variables';
+@forward 'fonts';
+@forward 'mixins';
+@forward 'utils';
+
+@import url('https://fonts.googleapis.com/css2?family=Roboto+Slab:wght@300;400;700&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Montserrat:wght@400;700&display=swap');
