+++ src/styles/abstracts/_utils.scss
@@ -1,0 +1,9 @@
+.d-none {
+  display: none;
+}
+
+.hidden {
+  position: absolute;
+  height: 0;
+  left: -10000px;
+}
