+++ src/styles/abstracts/_variables.scss
@@ -1,0 +1,21 @@
+// Primary Colors
+$black: #1e1e1e;
+$pinkish-grey: #c4c4c4;
+$warm-grey: #888;
+$black-two: #121212;
+$black: #000;
+$dodger-blue: #379cf6;
+$medium-blue: #307bbe;
+$black-5: rgba(0, 0, 0, 0.05);
+$azure: #0290ff;
+$pale-grey: #f5f6fa;
+$pale-blue: #4386bf;
+
+// Color Variables
+$primary: $azure;
+$primary-border-color: #e8e8e8;
+$bg-color: #fafafa;
+
+// White Space
+$gray: rgb(225, 226, 225);
+$white: rgb(255, 255, 255);
