+++ src/styles/index.scss
@@ -1,0 +1,3 @@
+// abstracts and base
+@import 'layout/layout';
+@import 'layout/typography';
