+++ src/styles/layout/_layout.scss
@@ -1,0 +1,73 @@
+@use '../abstracts/' as *;
+
+*,
+*::after,
+*::before {
+  box-sizing: inherit;
+  margin: 0;
+  padding: 0;
+}
+
+ul,
+li {
+  list-style: none;
+  margin: 0;
+  padding: 0;
+}
+
+html {
+  box-sizing: border-box;
+  font-size: 100%;
+}
+
+body {
+  font-size: 1.6rem;
+  color: rgb(0, 0, 0);
+  background-color: $bg-color;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+}
+
+a,
+a:visited,
+a:active,
+a:hover {
+  text-decoration: none;
+  font: inherit;
+  color: inherit;
+}
+
+button {
+  border: none;
+  cursor: pointer;
+  background-color: $primary;
+  color: $white;
+  border-radius: 4px;
+}
+
+hr {
+  border: $primary-border-color 1px solid;
+  margin-bottom: 2.6rem;
+}
+
+.page-section {
+  max-width: 1100px;
+  margin: 2.8rem auto;
+  padding: 0 1rem;
+}
+
+/*
+  ========================
+
+  Media Queries
+
+  ========================
+*/
+
+/* Large devices (laptops/desktops, 769px and up) */
+
+/* @media only screen and (min-width: 769px) {
+  .page-section {
+    padding: 0;
+  }
+} */
